Clean up: - Replace some C cast by C++ cast. - Don't use multi-line string. - Replace malloc by new []. - Replace sprintf by snprintf - Use some cUnit* in server.cpp to factorize code.
git-svn-id: svn://maxr.org/MAX/trunk@3264 c3bd5e88-ca0b-48cf-b972-da51c77c4c1b

+++ events.cpp
@@ -44,7 +44,7 @@ static std::string TakeScreenShot()
 	tmTime = localtime (&tTime);
 	strftime (timestr, sizeof (timestr), "%Y-%m-%d_%H%M%S", tmTime);
 	std::string screenshotfile;
-	short int counter = 0;
+	int counter = 0;
 	do
 	{
 		counter += 1;
